improve JSdocs. move config files to /config subfolder
