--- projects/storefrontlib/cms-structure/cms-lcp-context/lcp-cms-components.config.ts
@@ -1,17 +1,0 @@
-import { Injectable } from '@angular/core';
-import { Config } from '@spartacus/core';
-
-@Injectable({
-  providedIn: 'root',
-  useExisting: Config,
-})
-export abstract class LcpCmsComponentsConfig {
-  lcpCmsComponents?: {
-    idMarker?: string;
-    ids?: string[];
-  };
-}
-
-declare module '@spartacus/core' {
-  interface Config extends LcpCmsComponentsConfig {}
-}
